requirements.txt
python-telegram-bot[job-queue]==20.6
aiohttp==3.8.4
python-dotenv==1.0.0      
uvloop==0.17.0            
nest_asyncio
